requirements.txt
streamlit
tensorflow
pillow
numpy
streamlit-lottie
requests